requirements.txt
selenium>=4.23.1
webdriver-manager>=4.0.2
openpyxl>=3.1.5
requests>=2.31.0
python-dotenv>=1.0.1
